feat: create categories hook and add to selection
