+++ src/components/molecules/ChoiceCard/ChoiceCard.tsx
@@ -4,13 +4,13 @@ import { Text } from '../../atoms/Text/Text';
 
 import S from './ChoiceCard.module.scss';
 
-export const ChoiceCard = () => {
+export const ChoiceCard = ({ category }: { category: string }) => {
   return (
     <li className={S.card}>
       {/* <Picture size='small' image='choice' /> */}
       <Checkbox>
         <Text color='grey' size='default'>
-          sneakers
+          {category}
         </Text>
       </Checkbox>
     </li>
