+++ src/components/molecules/FilterCategories/FilterCategories.tsx
@@ -1,26 +1,9 @@
-import { useEffect } from 'react';
-import { useGetCategoriesQuery } from '../../../app/api/productsApi';
 import { Button } from '../../atoms/Button/Button';
 import { Title } from '../../atoms/Title/Title';
 import S from './FilterCategories.module.scss';
-import { useAppDispatch, useAppSelector } from '../../../hooks/reduxHooks';
-import { setCategories } from '../../../app/reducers/ProductsSlice';
+import { useCategories } from '../../../hooks/useCategories';
 export const FilterCategories = () => {
-  const dispatch = useAppDispatch();
-  const {
-    data: categoriesData,
-    isLoading,
-    isSuccess,
-  } = useGetCategoriesQuery(undefined);
-
-  useEffect(() => {
-    if (isSuccess) {
-      dispatch(setCategories(categoriesData));
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [categoriesData]);
-
-  const categories = useAppSelector(state => state.products.categories);
+  const { categories, isLoading } = useCategories();
 
   return (
     <div className={S.container}>
